x86/boot: Explain discard_initial_images() and fix PV initrd handling
discard_initial_images() frees the memory backing the modules in memory.  It
is called by dom0_construct_pv{,h}(), but obtains the module list by global
pointer which is why there is no apparent link with the initrd pointer.

Generally, module contents are copied into dom0 as it's being built, but
there's one case for PV dom0 where the initrd might be directly mapped instead
of copied.

dom0_construct_pv() does it's own ad-hoc freeing of the module in the copy
case, then unconditionally sets the size to 0.  This only functions because
init_domheap_pages(x, x) happens to be a nop.

Alter discard_initial_images() to explicitly check for an ignored module, and
explain what's going on.  This is more robust and will allow for fixing other
problems with module handling.  Delete the ad-hoc freeing().

Signed-off-by: Andrew Cooper <andrew.cooper3@citrix.com>
---
CC: Jan Beulich <JBeulich@suse.com>
CC: Roger Pau Monné <roger.pau@citrix.com>