xen/README: add compiler and binutils versions for RISC-V64
This patch doesn't represent a strict lower bound for GCC and
GNU Binutils; rather, these versions are specifically employed by
the Xen RISC-V container and are anticipated to undergo continuous
testing. Older GCC and GNU Binutils would work,
but this is not a guarantee.

While it is feasible to utilize Clang, it's important to note that,
currently, there is no Xen RISC-V CI job in place to verify the
seamless functioning of the build with Clang.

--
Changes in V5-V12:
 - Nothing changed. Only rebase.

Signed-off-by: Oleksii Kurochko <oleksii.kurochko@gmail.com>